`have_been_called_with` is not 'strict', but 'at least'.
This means that, when not negated, the matcher will pass if and
only if the spy was called with at least the specified arguments,
even if there were other arguments. But all specified arguments
must have been passed.

When negated, the meaning stays the same: the matcher will pass
if and only if the spy was not called with at least the specified
arguments, in other words, when at least one of the specified
arguments was missing from the call.